requirements.txt
fastapi
uvicorn
gunicorn
supabase
google-generativeai
python-dotenv
